requirements.txt
numpy
pandas
scikit-learn
xgboost
pyyaml
aqme
openbabel
rdkit
mordred